+++ master/buildbot/util/__init__.py
@@ -248,10 +248,10 @@ def human_readable_delta(start, end):
     if delta.days > 0:
         result.append('%d days' % (delta.days,))
     if delta.seconds > 0:
-        hours = delta.seconds / 3600
+        hours = int(delta.seconds / 3600)
         if hours > 0:
             result.append('%d hours' % (hours,))
-        minutes = (delta.seconds - hours * 3600) / 60
+        minutes = int((delta.seconds - hours * 3600) / 60)
         if minutes:
             result.append('%d minutes' % (minutes,))
         seconds = delta.seconds % 60
